+++ ProjectBj.BusinessLogic/Mappers/StandGameViewMapper.cs
@@ -39,7 +39,7 @@ namespace ProjectBj.BusinessLogic.Mappers
                 {
                     Suit = card.Suit.ToString(),
                     Rank = EnumHelper.GetCardRankName(card.Rank),
-                    RankValue = (int)card.Rank
+                    RankValue = card.Rank
                 };
                 cardStandGameViewItems.Add(cardStandGameViewItem);
             }
